Module google_datamigration1::api

Structs

AuditConfig
Specifies the audit configuration for a service. The configuration determines which permission types are logged, and what identities, if any, are exempted from logging. An AuditConfig must have one or more AuditLogConfigs. If there are AuditConfigs for both allServices and a specific service, the union of the two AuditConfigs is used for that service: the log_types specified in each AuditConfig are enabled, and the exempted_members in each AuditLogConfig are exempted. Example Policy with multiple AuditConfigs: { "audit_configs": [ { "service": "allServices", "audit_log_configs": [ { "log_type": "DATA_READ", "exempted_members": [ "user:jose@example.com" ] }, { "log_type": "DATA_WRITE" }, { "log_type": "ADMIN_READ" } ] }, { "service": "sampleservice.googleapis.com", "audit_log_configs": [ { "log_type": "DATA_READ" }, { "log_type": "DATA_WRITE", "exempted_members": [ "user:aliya@example.com" ] } ] } ] } For sampleservice, this policy enables DATA_READ, DATA_WRITE and ADMIN_READ logging. It also exempts jose@example.com from DATA_READ logging, and aliya@example.com from DATA_WRITE logging.

AuditLogConfig
Provides the configuration for logging a type of permissions. Example: { "audit_log_configs": [ { "log_type": "DATA_READ", "exempted_members": [ "user:jose@example.com" ] }, { "log_type": "DATA_WRITE" } ] } This enables 'DATA_READ' and 'DATA_WRITE' logging, while exempting jose@example.com from DATA_READ logging.

Binding
Associates members with a role.

CancelOperationRequest
The request message for Operations.CancelOperation.

CloudSqlConnectionProfile
Specifies required connection parameters, and, optionally, the parameters required to create a Cloud SQL destination database instance.

CloudSqlSettings
Settings for creating a Cloud SQL database instance.

ConnectionProfile
A connection profile definition.

DatabaseMigrationService
Central instance to access all DatabaseMigrationService related resource activities

DatabaseType
A message defining the database engine and provider.

Empty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

Expr
Represents a textual expression in the Common Expression Language (CEL) syntax. CEL is a C-like expression language. The syntax and semantics of CEL are documented at https://github.com/google/cel-spec. Example (Comparison): title: "Summary size limit" description: "Determines if a summary is less than 100 chars" expression: "document.summary.size() < 100" Example (Equality): title: "Requestor is owner" description: "Determines if requestor is the document owner" expression: "document.owner == request.auth.claims.email" Example (Logic): title: "Public documents" description: "Determine whether the document should be publicly visible" expression: "document.type != 'private' && document.type != 'internal'" Example (Data Manipulation): title: "Notification string" description: "Create a notification string with a timestamp." expression: "'New message received at ' + string(document.create_time)" The exact variables and functions that may be referenced within an expression are determined by the service that evaluates it. See the service documentation for additional information.

GenerateSshScriptRequest
Request message for 'GenerateSshScript' request.

ListConnectionProfilesResponse
Response message for 'ListConnectionProfiles' request.

ListLocationsResponse
The response message for Locations.ListLocations.

ListMigrationJobsResponse
Response message for 'ListMigrationJobs' request.

ListOperationsResponse
The response message for Operations.ListOperations.

Location
A resource that represents Google Cloud Platform location.

MigrationJob
Represents a Database Migration Service migration job object.

MySqlConnectionProfile
Specifies connection parameters required specifically for MySQL databases.

Operation
This resource represents a long-running operation that is the result of a network API call.

Policy
An Identity and Access Management (IAM) policy, which specifies access controls for Google Cloud resources. A Policy is a collection of bindings. A binding binds one or more members to a single role. Members can be user accounts, service accounts, Google groups, and domains (such as G Suite). A role is a named list of permissions; each role can be an IAM predefined role or a user-created custom role. For some types of Google Cloud resources, a binding can also specify a condition, which is a logical expression that allows access to a resource only if the expression evaluates to true. A condition can add constraints based on attributes of the request, the resource, or both. To learn which resources support conditions in their IAM policies, see the IAM documentation. JSON example: { "bindings": [ { "role": "roles/resourcemanager.organizationAdmin", "members": [ "user:mike@example.com", "group:admins@example.com", "domain:google.com", "serviceAccount:my-project-id@appspot.gserviceaccount.com" ] }, { "role": "roles/resourcemanager.organizationViewer", "members": [ "user:eve@example.com" ], "condition": { "title": "expirable access", "description": "Does not grant access after Sep 2020", "expression": "request.time < timestamp('2020-10-01T00:00:00.000Z')", } } ], "etag": "BwWWja0YfJA=", "version": 3 } YAML example: bindings: - members: - user:mike@example.com - group:admins@example.com - domain:google.com - serviceAccount:my-project-id@appspot.gserviceaccount.com role: roles/resourcemanager.organizationAdmin - members: - user:eve@example.com role: roles/resourcemanager.organizationViewer condition: title: expirable access description: Does not grant access after Sep 2020 expression: request.time < timestamp('2020-10-01T00:00:00.000Z') - etag: BwWWja0YfJA= - version: 3 For a description of IAM and its features, see the IAM documentation.

PostgreSqlConnectionProfile
Specifies connection parameters required specifically for PostgreSQL databases.

ProjectLocationConnectionProfileCreateCall
Creates a new connection profile in a given project and location.

ProjectLocationConnectionProfileDeleteCall
Deletes a single Database Migration Service connection profile. A connection profile can only be deleted if it is not in use by any active migration jobs.

ProjectLocationConnectionProfileGetCall
Gets details of a single connection profile.

ProjectLocationConnectionProfileGetIamPolicyCall
Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

ProjectLocationConnectionProfileListCall
Retrieve a list of all connection profiles in a given project and location.

ProjectLocationConnectionProfilePatchCall
Update the configuration of a single connection profile.

ProjectLocationConnectionProfileSetIamPolicyCall
Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

ProjectLocationConnectionProfileTestIamPermissionCall
Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

ProjectLocationGetCall
Gets information about a location.

ProjectLocationListCall
Lists information about the supported locations for this service.

ProjectLocationMigrationJobCreateCall
Creates a new migration job in a given project and location.

ProjectLocationMigrationJobDeleteCall
Deletes a single migration job.

ProjectLocationMigrationJobGenerateSshScriptCall
Generate a SSH configuration script to configure the reverse SSH connectivity.

ProjectLocationMigrationJobGetCall
Gets details of a single migration job.

ProjectLocationMigrationJobGetIamPolicyCall
Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

ProjectLocationMigrationJobListCall
Lists migration jobs in a given project and location.

ProjectLocationMigrationJobPatchCall
Updates the parameters of a single migration job.

ProjectLocationMigrationJobPromoteCall
Promote a migration job, stopping replication to the destination and promoting the destination to be a standalone database.

ProjectLocationMigrationJobRestartCall
Restart a stopped or failed migration job, resetting the destination instance to its original state and starting the migration process from scratch.

ProjectLocationMigrationJobResumeCall
Resume a migration job that is currently stopped and is resumable (was stopped during CDC phase).

ProjectLocationMigrationJobSetIamPolicyCall
Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

ProjectLocationMigrationJobStartCall
Start an already created migration job.

ProjectLocationMigrationJobStopCall
Stops a running migration job.

ProjectLocationMigrationJobTestIamPermissionCall
Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

ProjectLocationMigrationJobVerifyCall
Verify a migration job, making sure the destination can reach the source and that all configuration and prerequisites are met.

ProjectLocationOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

ProjectLocationOperationDeleteCall
Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

ProjectLocationOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectLocationOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the DatabaseMigrationService hub.

PromoteMigrationJobRequest
Request message for 'PromoteMigrationJob' request.

RestartMigrationJobRequest
Request message for 'RestartMigrationJob' request.

ResumeMigrationJobRequest
Request message for 'ResumeMigrationJob' request.

ReverseSshConnectivity
The details needed to configure a reverse SSH tunnel between the source and destination databases. These details will be used when calling the generateSshScript method (see https://cloud.google.com/database-migration/docs/reference/rest/v1/projects.locations.migrationJobs/generateSshScript) to produce the script that will help set up the reverse SSH tunnel, and to set up the VPC peering between the Cloud SQL private network and the VPC.

SetIamPolicyRequest
Request message for SetIamPolicy method.

SqlAclEntry
An entry for an Access Control list.

SqlIpConfig
IP Management configuration.

SshScript
Response message for 'GenerateSshScript' request.

SslConfig
SSL configuration information.

StartMigrationJobRequest
Request message for 'StartMigrationJob' request.

StaticIpConnectivity
The source database will allow incoming connections from the destination database's public IP. You can retrieve the Cloud SQL instance's public IP from the Cloud SQL console or using Cloud SQL APIs. No additional configuration is required.

Status
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

StopMigrationJobRequest
Request message for 'StopMigrationJob' request.

TestIamPermissionsRequest
Request message for TestIamPermissions method.

TestIamPermissionsResponse
Response message for TestIamPermissions method.

VerifyMigrationJobRequest
Request message for 'VerifyMigrationJob' request.

VmCreationConfig
VM creation configuration message

VmSelectionConfig
VM selection configuration message

VpcPeeringConnectivity
The details of the VPC where the source database is located in Google Cloud. We will use this information to set up the VPC peering connection between Cloud SQL and this VPC.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.