Struct google_datamigration1::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the DatabaseMigrationService hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_datamigration1 as datamigration1;
 
use std::default::Default;
use oauth2;
use datamigration1::DatabaseMigrationService;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = DatabaseMigrationService::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `locations_connection_profiles_create(...)`, `locations_connection_profiles_delete(...)`, `locations_connection_profiles_get(...)`, `locations_connection_profiles_get_iam_policy(...)`, `locations_connection_profiles_list(...)`, `locations_connection_profiles_patch(...)`, `locations_connection_profiles_set_iam_policy(...)`, `locations_connection_profiles_test_iam_permissions(...)`, `locations_get(...)`, `locations_list(...)`, `locations_migration_jobs_create(...)`, `locations_migration_jobs_delete(...)`, `locations_migration_jobs_generate_ssh_script(...)`, `locations_migration_jobs_get(...)`, `locations_migration_jobs_get_iam_policy(...)`, `locations_migration_jobs_list(...)`, `locations_migration_jobs_patch(...)`, `locations_migration_jobs_promote(...)`, `locations_migration_jobs_restart(...)`, `locations_migration_jobs_resume(...)`, `locations_migration_jobs_set_iam_policy(...)`, `locations_migration_jobs_start(...)`, `locations_migration_jobs_stop(...)`, `locations_migration_jobs_test_iam_permissions(...)`, `locations_migration_jobs_verify(...)`, `locations_operations_cancel(...)`, `locations_operations_delete(...)`, `locations_operations_get(...)` and `locations_operations_list(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn locations_connection_profiles_create(
&self,
request: ConnectionProfile,
parent: &str
) -> ProjectLocationConnectionProfileCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new connection profile in a given project and location.

Arguments

request - No description provided.

parent - Required. The parent, which owns this collection of connection profiles.

pub fn locations_connection_profiles_delete(
&self,
name: &str
) -> ProjectLocationConnectionProfileDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single Database Migration Service connection profile. A connection profile can only be deleted if it is not in use by any active migration jobs.

Arguments

name - Required. Name of the connection profile resource to delete.

pub fn locations_connection_profiles_get(
&self,
name: &str
) -> ProjectLocationConnectionProfileGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single connection profile.

Arguments

name - Required. Name of the connection profile resource to get.

pub fn locations_connection_profiles_get_iam_policy(
&self,
resource: &str
) -> ProjectLocationConnectionProfileGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_connection_profiles_list(
&self,
parent: &str
) -> ProjectLocationConnectionProfileListCall<'a, C>

Create a builder to help you perform the following task:

Retrieve a list of all connection profiles in a given project and location.

Arguments

parent - Required. The parent, which owns this collection of connection profiles.

pub fn locations_connection_profiles_patch(
&self,
request: ConnectionProfile,
name: &str
) -> ProjectLocationConnectionProfilePatchCall<'a, C>

Create a builder to help you perform the following task:

Update the configuration of a single connection profile.

Arguments

request - No description provided.

name - The name of this connection profile resource in the form of projects/{project}/locations/{location}/instances/{instance}.

pub fn locations_connection_profiles_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationConnectionProfileSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_connection_profiles_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationConnectionProfileTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_migration_jobs_create(
&self,
request: MigrationJob,
parent: &str
) -> ProjectLocationMigrationJobCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a new migration job in a given project and location.

Arguments

request - No description provided.

parent - Required. The parent, which owns this collection of migration jobs.

pub fn locations_migration_jobs_delete(
&self,
name: &str
) -> ProjectLocationMigrationJobDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a single migration job.

Arguments

name - Required. Name of the migration job resource to delete.

pub fn locations_migration_jobs_generate_ssh_script(
&self,
request: GenerateSshScriptRequest,
migration_job: &str
) -> ProjectLocationMigrationJobGenerateSshScriptCall<'a, C>

Create a builder to help you perform the following task:

Generate a SSH configuration script to configure the reverse SSH connectivity.

Arguments

request - No description provided.

migrationJob - Name of the migration job resource to generate the SSH script.

pub fn locations_migration_jobs_get(
&self,
name: &str
) -> ProjectLocationMigrationJobGetCall<'a, C>

Create a builder to help you perform the following task:

Gets details of a single migration job.

Arguments

name - Required. Name of the migration job resource to get.

pub fn locations_migration_jobs_get_iam_policy(
&self,
resource: &str
) -> ProjectLocationMigrationJobGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. Returns an empty policy if the resource exists and does not have a policy set.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_migration_jobs_list(
&self,
parent: &str
) -> ProjectLocationMigrationJobListCall<'a, C>

Create a builder to help you perform the following task:

Lists migration jobs in a given project and location.

Arguments

parent - Required. The parent, which owns this collection of migrationJobs.

pub fn locations_migration_jobs_patch(
&self,
request: MigrationJob,
name: &str
) -> ProjectLocationMigrationJobPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates the parameters of a single migration job.

Arguments

request - No description provided.

name - The name (URI) of this migration job resource, in the form of: projects/{project}/locations/{location}/instances/{instance}.

pub fn locations_migration_jobs_promote(
&self,
request: PromoteMigrationJobRequest,
name: &str
) -> ProjectLocationMigrationJobPromoteCall<'a, C>

Create a builder to help you perform the following task:

Promote a migration job, stopping replication to the destination and promoting the destination to be a standalone database.

Arguments

request - No description provided.

name - Name of the migration job resource to promote.

pub fn locations_migration_jobs_restart(
&self,
request: RestartMigrationJobRequest,
name: &str
) -> ProjectLocationMigrationJobRestartCall<'a, C>

Create a builder to help you perform the following task:

Restart a stopped or failed migration job, resetting the destination instance to its original state and starting the migration process from scratch.

Arguments

request - No description provided.

name - Name of the migration job resource to restart.

pub fn locations_migration_jobs_resume(
&self,
request: ResumeMigrationJobRequest,
name: &str
) -> ProjectLocationMigrationJobResumeCall<'a, C>

Create a builder to help you perform the following task:

Resume a migration job that is currently stopped and is resumable (was stopped during CDC phase).

Arguments

request - No description provided.

name - Name of the migration job resource to resume.

pub fn locations_migration_jobs_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationMigrationJobSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy on the specified resource. Replaces any existing policy. Can return NOT_FOUND, INVALID_ARGUMENT, and PERMISSION_DENIED errors.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_migration_jobs_start(
&self,
request: StartMigrationJobRequest,
name: &str
) -> ProjectLocationMigrationJobStartCall<'a, C>

Create a builder to help you perform the following task:

Start an already created migration job.

Arguments

request - No description provided.

name - Name of the migration job resource to start.

pub fn locations_migration_jobs_stop(
&self,
request: StopMigrationJobRequest,
name: &str
) -> ProjectLocationMigrationJobStopCall<'a, C>

Create a builder to help you perform the following task:

Stops a running migration job.

Arguments

request - No description provided.

name - Name of the migration job resource to stop.

pub fn locations_migration_jobs_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationMigrationJobTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns permissions that a caller has on the specified resource. If the resource does not exist, this will return an empty set of permissions, not a NOT_FOUND error. Note: This operation is designed to be used for building permission-aware UIs and command-line tools, not for authorization checking. This operation may "fail open" without warning.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_migration_jobs_verify(
&self,
request: VerifyMigrationJobRequest,
name: &str
) -> ProjectLocationMigrationJobVerifyCall<'a, C>

Create a builder to help you perform the following task:

Verify a migration job, making sure the destination can reach the source and that all configuration and prerequisites are met.

Arguments

request - No description provided.

name - Name of the migration job resource to verify.

pub fn locations_operations_cancel(
&self,
request: CancelOperationRequest,
name: &str
) -> ProjectLocationOperationCancelCall<'a, C>

Create a builder to help you perform the following task:

Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

Arguments

request - No description provided.

name - The name of the operation resource to be cancelled.

pub fn locations_operations_delete(
&self,
name: &str
) -> ProjectLocationOperationDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

Arguments

name - The name of the operation resource to be deleted.

pub fn locations_operations_get(
&self,
name: &str
) -> ProjectLocationOperationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

Arguments

name - The name of the operation resource.

pub fn locations_operations_list(
&self,
name: &str
) -> ProjectLocationOperationListCall<'a, C>

Create a builder to help you perform the following task:

Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

Arguments

name - The name of the operation's parent resource.

pub fn locations_get(&self, name: &str) -> ProjectLocationGetCall<'a, C>

Create a builder to help you perform the following task:

Gets information about a location.

Arguments

name - Resource name for the location.

pub fn locations_list(&self, name: &str) -> ProjectLocationListCall<'a, C>

Create a builder to help you perform the following task:

Lists information about the supported locations for this service.

Arguments

name - The resource that owns the locations collection, if applicable.